Only send 3 gratuitous ARP packets after switching to ACTIVATED state
gARP goal is to enfore a change of ARP tables in the infrastructure after a failover happens,
when the ACTIVATED node for an IP is changing. Thus it should not be sent continously but only
when a change happen.

That's what is describe here

https://www.practicalnetworking.net/series/arp/gratuitous-arp/

And applied in the VRRP protocol for instance

https://tools.ietf.org/html/rfc3768

Thus we're now only sending them at that moment, and then servers will answer normal ARPs requests afterwards, the three gARP should be enough to trigger the table update everywhere, it's what is done by keepalived

Fixes #79

The issue is fixed because it was related to the point that the service was sending continously too many gARP at the same time, now the number of gARP has been decreased
and some randomization has been added to avoid all gARPs to be sent at the same time